requirements.txt
python-telegram-bot[job-queue]==21.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
schedule==1.2.2